+++ app/Http/Controllers/Dashboard/RoleController.php
@@ -6,14 +6,24 @@ use App\Http\Controllers\Controller;
 use App\Models\Role;
 use App\Models\RoleAbility;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
+
+   public function __construct()
+    {
+        // $this->authorizeResource(Role::class,'id');//if parameter is id in controller
+        // $this->authorizeResource(Role::class,'role');//if parameter is object in controller //if no policy for this model it will return false that no abilities fot it 
+    }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        $this->authorize('viewAny',Role::class);
+        // Gate::authorize('role.view');
+
         $roles = Role::paginate(10);
         return view('dashboard.roles.index', compact('roles'));
     }
@@ -23,6 +33,10 @@ class RoleController extends Controller
      */
     public function create()
     {
+        $this->authorize('create',Role::class);
+
+        // Gate::authorize('role.create');
+
         $allow = 'allow';
 
         $role = new Role();
@@ -34,6 +48,8 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create',Role::class);
+
         $request->validate([
             'name' => 'required|string|max:255',
             'abilities' => 'required|array'
@@ -68,9 +84,11 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
-
+        $this->authorize('update',$role);
+        // Gate::authorize('role.update');
+        $allow = 'allow';
         $roleAbility = $role->roleAbility()->pluck('type', 'ability')->toArray();
-        return view('dashboard.roles.edit', compact('roleAbility', 'role'));
+        return view('dashboard.roles.edit', compact('roleAbility', 'role','allow'));
     }
 
     /**
@@ -78,6 +96,8 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
+        $this->authorize('update',$role);
+
         $request->validate([
             'name' => 'required|string|max:255|unique:roles,name,' . $role->id,
             'abilities' => 'required|array'
@@ -107,6 +127,10 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        $this->authorize('delete',$role);
+
+        // Gate::authorize('role.delete');
+
         $role->delete();
         return redirect()->route('roles.index')
             ->with('success', 'Data deleted successfully.');
